refactor(home): move playerProfileUrl into utilities
It's a dependency-free string helper, not a component — utilities/ is where
those already live (GameConfigHelpers, ReplaySpeedMultiplier).
